Algebra.Internal

Documentation

data (a :: k) :~: (b :: k) :: forall k. k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: a :~: a

withRefl :: (a :~: b) -> ((a ~ b) -> r) -> r

module Data.Proxy

class KnownNat (n :: Nat)

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: 4.7.0.0

Minimal complete definition

natSing

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: 4.7.0.0

Minimal complete definition

symbolSing

data Nat

(Kind) This is the kind of type-level natural numbers.

data Symbol

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

type family (a :: Nat) ^ (b :: Nat) :: Nat where ... infixr 8

Exponentiation of type-level naturals.

Since: 4.7.0.0

type family (a :: Nat) <=? (b :: Nat) :: Bool where ... infix 4

Comparison of type-level naturals, as a function. NOTE: The functionality for this function should be subsumed by CmpNat, so this might go away in the future. Please let us know, if you encounter discrepancies between the two.

type family CmpSymbol (a :: Symbol) (b :: Symbol) :: Ordering where ...

Comparison of type-level symbols, as a function.

Since: 4.7.0.0

type family CmpNat (a :: Nat) (b :: Nat) :: Ordering where ...

Comparison of type-level naturals, as a function.

Since: 4.7.0.0

type family Div (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Division (round down) of natural numbers. Div x 0 is undefined (i.e., it cannot be reduced).

Since: 4.11.0.0

type family Mod (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Modulus of natural numbers. Mod x 0 is undefined (i.e., it cannot be reduced).

Since: 4.11.0.0

type family Log2 (a :: Nat) :: Nat where ...

Log base 2 (round down) of natural numbers. Log 0 is undefined (i.e., it cannot be reduced).

Since: 4.11.0.0

type family TypeError (a :: ErrorMessage) :: b where ...

The type-level equivalent of error.

The polymorphic kind of this type allows it to be used in several settings. For instance, it can be used as a constraint, e.g. to provide a better error message for a non-existent instance,

-- in a context
instance TypeError (Text "Cannot Show functions." :$$:
                    Text "Perhaps there is a missing argument?")
      => Show (a -> b) where
    showsPrec = error "unreachable"

It can also be placed on the right-hand side of a type-level function to provide an error for an invalid case,

type family ByteSize x where
   ByteSize Word16   = 2
   ByteSize Word8    = 1
   ByteSize a        = TypeError (Text "The type " :<>: ShowType a :<>:
                                  Text " is not exportable.")

Since: 4.9.0.0

type family AppendSymbol (a :: Symbol) (b :: Symbol) :: Symbol where ...

Concatenation of type-level symbols.

Since: 4.10.0.0

sameSymbol :: (KnownSymbol a, KnownSymbol b) => Proxy a -> Proxy b -> Maybe (a :~: b)

We either get evidence that this function was instantiated with the same type-level symbols, or Nothing.

Since: 4.7.0.0

someSymbolVal :: String -> SomeSymbol

Convert a string into an unknown type-level symbol.

Since: 4.7.0.0

someNatVal :: Integer -> Maybe SomeNat

Convert an integer into an unknown type-level natural.

Since: 4.7.0.0

symbolVal' :: KnownSymbol n => Proxy# n -> String

Since: 4.8.0.0

natVal' :: KnownNat n => Proxy# n -> Integer

Since: 4.8.0.0

symbolVal :: KnownSymbol n => proxy n -> String

Since: 4.7.0.0

natVal :: KnownNat n => proxy n -> Integer

Since: 4.7.0.0

data SomeSymbol where

This type represents unknown type-level symbols.

Constructors

SomeSymbol :: SomeSymbol
Since: 4.7.0.0

data ErrorMessage where

A description of a custom type error.

Constructors

Text :: ErrorMessage
Show the text as is.

ShowType :: ErrorMessage
Pretty print the type. ShowType :: k -> ErrorMessage

(:<>:) :: ErrorMessage infixl 6
Put two pieces of error message next to each other.

(:$$:) :: ErrorMessage infixl 5
Stack two pieces of error message on top of each other.

sameNat :: (KnownNat a, KnownNat b) => Proxy a -> Proxy b -> Maybe (a :~: b)

We either get evidence that this function was instantiated with the same type-level numbers, or Nothing.

Since: 4.7.0.0

data SomeNat where

This type represents unknown type-level natural numbers.

Since: 4.10.0.0

Constructors

SomeNat :: SomeNat

withSingI :: Sing n -> (SingI n -> r) -> r

data family Sing (a :: k) :: *

class SingI (a :: k) where

Minimal complete definition

sing

Methods

sing :: Sing a

class SingKind k where

Minimal complete definition

fromSing, toSing

Associated Types

type Demote k = (r :: *) | r -> k

Methods

fromSing :: Sing a -> Demote k

toSing :: Demote k -> SomeSing k

data SomeSing k where

Constructors

SomeSing :: SomeSing k

class PEnum a

Associated Types

type Succ (arg :: a) :: a

type Pred (arg :: a) :: a

type ToEnum (arg :: Nat) :: a

type FromEnum (arg :: a) :: Nat

type EnumFromTo (arg :: a) (arg1 :: a) :: [a]

type EnumFromThenTo (arg :: a) (arg1 :: a) (arg2 :: a) :: [a]

class SEnum a where

Minimal complete definition

sToEnum, sFromEnum

Methods

sSucc :: Sing t -> Sing (Apply (SuccSym0 :: TyFun a a -> *) t)

sPred :: Sing t -> Sing (Apply (PredSym0 :: TyFun a a -> *) t)

sToEnum :: Sing t -> Sing (Apply (ToEnumSym0 :: TyFun Nat a -> *) t)

sFromEnum :: Sing t -> Sing (Apply (FromEnumSym0 :: TyFun a Nat -> *) t)

sEnumFromTo :: Sing t1 -> Sing t2 -> Sing (Apply (Apply (EnumFromToSym0 :: TyFun a (TyFun a [a] -> Type) -> *) t1) t2)

sEnumFromThenTo :: Sing t1 -> Sing t2 -> Sing t3 -> Sing (Apply (Apply (Apply (EnumFromThenToSym0 :: TyFun a (TyFun a (TyFun a [a] -> Type) -> Type) -> *) t1) t2) t3)

data (/=@#@$) (l :: TyFun a6989586621679557194 (TyFun a6989586621679557194 Bool -> Type)) :: forall a6989586621679557194. TyFun a6989586621679557194 (TyFun a6989586621679557194 Bool -> Type) -> *

data (l :: a6989586621679557194) /=@#@$$ (l1 :: TyFun a6989586621679557194 Bool) :: forall a6989586621679557194. a6989586621679557194 -> TyFun a6989586621679557194 Bool -> *

type (/=@#@$$$) (t :: a6989586621679557194) (t1 :: a6989586621679557194) = t /= t1

data (==@#@$) (l :: TyFun a6989586621679557194 (TyFun a6989586621679557194 Bool -> Type)) :: forall a6989586621679557194. TyFun a6989586621679557194 (TyFun a6989586621679557194 Bool -> Type) -> *

data (l :: a6989586621679557194) ==@#@$$ (l1 :: TyFun a6989586621679557194 Bool) :: forall a6989586621679557194. a6989586621679557194 -> TyFun a6989586621679557194 Bool -> *

type (==@#@$$$) (t :: a6989586621679557194) (t1 :: a6989586621679557194) = t == t1

type family (x :: a) == (y :: a) :: Bool

type family (x :: a) /= (y :: a) :: Bool

(%/=) :: SEq k => Sing a -> Sing b -> Sing (a /= b)

(%==) :: SEq k => Sing a -> Sing b -> Sing (a == b)

type EQSym0 = EQ

type GTSym0 = GT

type LTSym0 = LT

data (*@#@$) (l :: TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type)) :: forall a6989586621679686791. TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type) -> *

data (l :: a6989586621679686791) *@#@$$ (l1 :: TyFun a6989586621679686791 a6989586621679686791) :: forall a6989586621679686791. a6989586621679686791 -> TyFun a6989586621679686791 a6989586621679686791 -> *

type (*@#@$$$) (t :: a6989586621679686791) (t1 :: a6989586621679686791) = t * t1

data (+@#@$) (l :: TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type)) :: forall a6989586621679686791. TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type) -> *

data (l :: a6989586621679686791) +@#@$$ (l1 :: TyFun a6989586621679686791 a6989586621679686791) :: forall a6989586621679686791. a6989586621679686791 -> TyFun a6989586621679686791 a6989586621679686791 -> *

type (+@#@$$$) (t :: a6989586621679686791) (t1 :: a6989586621679686791) = t + t1

data (-@#@$) (l :: TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type)) :: forall a6989586621679686791. TyFun a6989586621679686791 (TyFun a6989586621679686791 a6989586621679686791 -> Type) -> *

data (l :: a6989586621679686791) -@#@$$ (l1 :: TyFun a6989586621679686791 a6989586621679686791) :: forall a6989586621679686791. a6989586621679686791 -> TyFun a6989586621679686791 a6989586621679686791 -> *

type (-@#@$$$) (t :: a6989586621679686791) (t1 :: a6989586621679686791) = t - t1

data FromIntegerSym0 (l :: TyFun Nat a6989586621679686791) :: forall a6989586621679686791. TyFun Nat a6989586621679686791 -> *

type FromIntegerSym1 (t :: Nat) = (FromInteger t :: k)

class PNum a

Associated Types

type (arg :: a) + (arg1 :: a) :: a

type (arg :: a) - (arg1 :: a) :: a

type (arg :: a) * (arg1 :: a) :: a

type Negate (arg :: a) :: a

type Abs (arg :: a) :: a

type Signum (arg :: a) :: a

type FromInteger (arg :: Nat) :: a

class SNum a where

Minimal complete definition

(%+), (%*), sAbs, sSignum, sFromInteger

Methods

(%+) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((+@#@$) :: TyFun a (TyFun a a -> Type) -> *) t1) t2)

(%-) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((-@#@$) :: TyFun a (TyFun a a -> Type) -> *) t1) t2)

(%*) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((*@#@$) :: TyFun a (TyFun a a -> Type) -> *) t1) t2)

sNegate :: Sing t -> Sing (Apply (NegateSym0 :: TyFun a a -> *) t)

sAbs :: Sing t -> Sing (Apply (AbsSym0 :: TyFun a a -> *) t)

sSignum :: Sing t -> Sing (Apply (SignumSym0 :: TyFun a a -> *) t)

sFromInteger :: Sing t -> Sing (Apply (FromIntegerSym0 :: TyFun Nat a -> *) t)

data (<=@#@$) (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type) -> *

data (l :: a6989586621679568425) <=@#@$$ (l1 :: TyFun a6989586621679568425 Bool) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 Bool -> *

type (<=@#@$$$) (t :: a6989586621679568425) (t1 :: a6989586621679568425) = t <= t1

data (<@#@$) (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type) -> *

data (l :: a6989586621679568425) <@#@$$ (l1 :: TyFun a6989586621679568425 Bool) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 Bool -> *

type (<@#@$$$) (t :: a6989586621679568425) (t1 :: a6989586621679568425) = t < t1

data (>=@#@$) (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type) -> *

data (l :: a6989586621679568425) >=@#@$$ (l1 :: TyFun a6989586621679568425 Bool) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 Bool -> *

type (>=@#@$$$) (t :: a6989586621679568425) (t1 :: a6989586621679568425) = t >= t1

data (>@#@$) (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 Bool -> Type) -> *

data (l :: a6989586621679568425) >@#@$$ (l1 :: TyFun a6989586621679568425 Bool) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 Bool -> *

type (>@#@$$$) (t :: a6989586621679568425) (t1 :: a6989586621679568425) = t > t1

data CompareSym0 (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 Ordering -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 Ordering -> Type) -> *

data CompareSym1 (l :: a6989586621679568425) (l1 :: TyFun a6989586621679568425 Ordering) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 Ordering -> *

type CompareSym2 (t :: a6989586621679568425) (t1 :: a6989586621679568425) = Compare t t1

data MaxSym0 (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 a6989586621679568425 -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 a6989586621679568425 -> Type) -> *

data MaxSym1 (l :: a6989586621679568425) (l1 :: TyFun a6989586621679568425 a6989586621679568425) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 a6989586621679568425 -> *

type MaxSym2 (t :: a6989586621679568425) (t1 :: a6989586621679568425) = Max t t1

data MinSym0 (l :: TyFun a6989586621679568425 (TyFun a6989586621679568425 a6989586621679568425 -> Type)) :: forall a6989586621679568425. TyFun a6989586621679568425 (TyFun a6989586621679568425 a6989586621679568425 -> Type) -> *

data MinSym1 (l :: a6989586621679568425) (l1 :: TyFun a6989586621679568425 a6989586621679568425) :: forall a6989586621679568425. a6989586621679568425 -> TyFun a6989586621679568425 a6989586621679568425 -> *

type MinSym2 (t :: a6989586621679568425) (t1 :: a6989586621679568425) = Min t t1

class PEq a => POrd a

Associated Types

type Compare (arg :: a) (arg1 :: a) :: Ordering

type (arg :: a) < (arg1 :: a) :: Bool

type (arg :: a) <= (arg1 :: a) :: Bool

type (arg :: a) > (arg1 :: a) :: Bool

type (arg :: a) >= (arg1 :: a) :: Bool

type Max (arg :: a) (arg1 :: a) :: a

type Min (arg :: a) (arg1 :: a) :: a

class SEq a => SOrd a where

Methods

sCompare :: Sing t1 -> Sing t2 -> Sing (Apply (Apply (CompareSym0 :: TyFun a (TyFun a Ordering -> Type) -> *) t1) t2)

(%<) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((<@#@$) :: TyFun a (TyFun a Bool -> Type) -> *) t1) t2)

(%<=) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((<=@#@$) :: TyFun a (TyFun a Bool -> Type) -> *) t1) t2)

(%>) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((>@#@$) :: TyFun a (TyFun a Bool -> Type) -> *) t1) t2)

(%>=) :: Sing t1 -> Sing t2 -> Sing (Apply (Apply ((>=@#@$) :: TyFun a (TyFun a Bool -> Type) -> *) t1) t2)

sMax :: Sing t1 -> Sing t2 -> Sing (Apply (Apply (MaxSym0 :: TyFun a (TyFun a a -> Type) -> *) t1) t2)

sMin :: Sing t1 -> Sing t2 -> Sing (Apply (Apply (MinSym0 :: TyFun a (TyFun a a -> Type) -> *) t1) t2)

withKnownNat :: Sing n -> (KnownNat n -> r) -> r

generate :: (ListLike (f a) a, HasOrdinal nat) => Sing n -> (Ordinal n -> a) -> Sized f n a

sIndex :: (HasOrdinal nat, ListLike (f c) c) => Ordinal n -> Sized f n c -> c

singleton :: ListLike (f a) a => a -> Sized f 1 a

unsafeFromList :: ListLike (f a) a => Sing n -> [a] -> Sized f n a

unsafeFromList' :: (SingI n, ListLike (f a) a) => [a] -> Sized f n a

zipWithSame :: (ListLike (f a) a, ListLike (f b) b, ListLike (f c) c) => (a -> b -> c) -> Sized f n a -> Sized f n b -> Sized f n c

pattern (:<) :: forall (f :: * -> *) (n :: Nat) a. ListLike (f a) a => forall (n1 :: Nat). (n ~ Succ n1, SingI n1) => a -> Sized f n1 a -> Sized f n a

pattern (:>) :: forall (f :: * -> *) (n :: Nat) a. ListLike (f a) a => forall (n1 :: Nat). (n ~ Succ n1, SingI n1) => Sized f n1 a -> a -> Sized f n a

pattern NilL :: forall (f :: * -> *) (n :: Nat) a. ListLike (f a) a => n ~ 0 => Sized f n a

pattern NilR :: forall (f :: * -> *) (n :: Nat) a. (ListLike (f a) a, SingI n) => n ~ 0 => Sized f n a

type family Zero nat :: nat where ...

Equations

Zero nat = (FromInteger 0 :: nat)

mkSNatQQ :: TypeQ -> QuasiQuoter

pattern Succ :: forall nat (n :: nat). IsPeano nat => forall (n1 :: nat). n ~ Succ n1 => Sing n1 -> Sing n

pattern Zero :: forall nat (n :: nat). IsPeano nat => n ~ Zero nat => Sing n

minusCong :: (n :~: m) -> (l :~: k) -> (n - l) :~: (m - k)

minusCongL :: (n :~: m) -> Sing k -> (n - k) :~: (m - k)

minusCongR :: Sing k -> (n :~: m) -> (k - n) :~: (k - m)

multCong :: (n :~: m) -> (l :~: k) -> (n * l) :~: (m * k)

multCongL :: (n :~: m) -> Sing k -> (n * k) :~: (m * k)

multCongR :: Sing k -> (n :~: m) -> (k * n) :~: (k * m)

plusCong :: (n :~: m) -> (n' :~: m') -> (n + n') :~: (m + m')

plusCongL :: (n :~: m) -> Sing k -> (n + k) :~: (m + k)

plusCongR :: Sing k -> (n :~: m) -> (k + n) :~: (k + m)

sOne :: SNum nat => Sing (One nat)

sZero :: SNum nat => Sing (Zero nat)

succCong :: (n :~: m) -> S n :~: S m

(%-.) :: PeanoOrder nat => Sing n -> Sing m -> Sing (n -. m)

coerceLeqL :: IsPeano nat => (n :~: m) -> Sing l -> IsTrue (n <= l) -> IsTrue (m <= l)

coerceLeqR :: IsPeano nat => Sing l -> (n :~: m) -> IsTrue (l <= n) -> IsTrue (l <= m)

minPlusTruncMinus :: PeanoOrder nat => Sing n -> Sing m -> (Min n m + (n -. m)) :~: n

sFlipOrdering :: Sing t -> Sing (Apply FlipOrderingSym0 t)

sLeqCong :: (a2 :~: b) -> (c :~: d) -> (a2 <= c) :~: (b <= d)

sLeqCongL :: (a2 :~: b) -> Sing c -> (a2 <= c) :~: (b <= c)

sLeqCongR :: Sing a2 -> (b :~: c) -> (a2 <= b) :~: (a2 <= c)

truncMinusLeq :: PeanoOrder nat => Sing n -> Sing m -> IsTrue ((n -. m) <= n)

class (SDecide nat, SNum nat, SEnum nat, SingKind nat, SingKind nat) => IsPeano nat where

Minimal complete definition

succOneCong, succNonCyclic, predSucc, plusMinus, succInj, (plusZeroL, plusSuccL | plusZeroR, plusZeroL), (multZeroL, multSuccL | multZeroR, multSuccR), induction

Methods

succOneCong :: Succ (Zero nat) :~: One nat

succInj :: (Succ n :~: Succ m) -> n :~: m

succInj' :: proxy n -> proxy' m -> (Succ n :~: Succ m) -> n :~: m

succNonCyclic :: Sing n -> (Succ n :~: Zero nat) -> Void

induction :: p (Zero nat) -> (forall (n :: nat). Sing n -> p n -> p (S n)) -> Sing k -> p k

plusMinus :: Sing n -> Sing m -> ((n + m) - m) :~: n

plusMinus' :: Sing n -> Sing m -> ((n + m) - n) :~: m

plusZeroL :: Sing n -> (Zero nat + n) :~: n

plusSuccL :: Sing n -> Sing m -> (S n + m) :~: S (n + m)

plusZeroR :: Sing n -> (n + Zero nat) :~: n

plusSuccR :: Sing n -> Sing m -> (n + S m) :~: S (n + m)

plusComm :: Sing n -> Sing m -> (n + m) :~: (m + n)

plusAssoc :: Sing n -> Sing m -> Sing l -> ((n + m) + l) :~: (n + (m + l))

multZeroL :: Sing n -> (Zero nat * n) :~: Zero nat

multSuccL :: Sing n -> Sing m -> (S n * m) :~: ((n * m) + m)

multZeroR :: Sing n -> (n * Zero nat) :~: Zero nat

multSuccR :: Sing n -> Sing m -> (n * S m) :~: ((n * m) + n)

multComm :: Sing n -> Sing m -> (n * m) :~: (m * n)

multOneR :: Sing n -> (n * One nat) :~: n

multOneL :: Sing n -> (One nat * n) :~: n

plusMultDistrib :: Sing n -> Sing m -> Sing l -> ((n + m) * l) :~: ((n * l) + (m * l))

multPlusDistrib :: Sing n -> Sing m -> Sing l -> (n * (m + l)) :~: ((n * m) + (n * l))

minusNilpotent :: Sing n -> (n - n) :~: Zero nat

multAssoc :: Sing n -> Sing m -> Sing l -> ((n * m) * l) :~: (n * (m * l))

plusEqCancelL :: Sing n -> Sing m -> Sing l -> ((n + m) :~: (n + l)) -> m :~: l

plusEqCancelR :: Sing n -> Sing m -> Sing l -> ((n + l) :~: (m + l)) -> n :~: m

succAndPlusOneL :: Sing n -> Succ n :~: (One nat + n)

succAndPlusOneR :: Sing n -> Succ n :~: (n + One nat)

predSucc :: Sing n -> Pred (Succ n) :~: n

zeroOrSucc :: Sing n -> ZeroOrSucc n

plusEqZeroL :: Sing n -> Sing m -> ((n + m) :~: Zero nat) -> n :~: Zero nat

plusEqZeroR :: Sing n -> Sing m -> ((n + m) :~: Zero nat) -> m :~: Zero nat

predUnique :: Sing n -> Sing m -> (Succ n :~: m) -> n :~: Pred m

multEqSuccElimL :: Sing n -> Sing m -> Sing l -> ((n * m) :~: Succ l) -> n :~: Succ (Pred n)

multEqSuccElimR :: Sing n -> Sing m -> Sing l -> ((n * m) :~: Succ l) -> m :~: Succ (Pred m)

minusZero :: Sing n -> (n - Zero nat) :~: n

multEqCancelR :: Sing n -> Sing m -> Sing l -> ((n * Succ l) :~: (m * Succ l)) -> n :~: m

succPred :: Sing n -> ((n :~: Zero nat) -> Void) -> Succ (Pred n) :~: n

multEqCancelL :: Sing n -> Sing m -> Sing l -> ((Succ n * m) :~: (Succ n * l)) -> m :~: l

sPred' :: proxy n -> Sing (Succ n) -> Sing n

toNatural :: Sing n -> Natural

type family One nat :: nat where ...

Equations

One nat = (FromInteger 1 :: nat)

type S (n :: k) = Succ n

data ZeroOrSucc (n :: nat) :: forall nat. nat -> * where

Constructors

IsZero :: ZeroOrSucc (Zero nat)

IsSucc :: ZeroOrSucc (Succ n1)

type (-.) (n :: k) (m :: k) = Subt n m (m <= n)

data DiffNat (n :: k) (m :: k) :: forall k. k -> k -> * where

Constructors

DiffNat :: DiffNat n (n + m1)

type family FlipOrdering (a :: Ordering) :: Ordering where ...

Equations

FlipOrdering EQ = EQSym0

FlipOrdering LT = GTSym0

FlipOrdering GT = LTSym0

data LeqView (n :: nat) (m :: nat) :: forall nat. nat -> nat -> * where

Constructors

LeqZero :: LeqView (Zero nat) m

LeqSucc :: LeqView (Succ n1) (Succ m1)

class (SOrd nat, IsPeano nat) => PeanoOrder nat where

Minimal complete definition

(succLeqToLT, cmpZero, leqRefl | leqZero, leqSucc, viewLeq | leqWitness, leqStep), eqlCmpEQ, ltToLeq, eqToRefl, flipCompare, leqToCmp, leqReversed, lneqSuccLeq, lneqReversed, (leqToMin, geqToMin | minLeqL, minLeqR, minLargest), (leqToMax, geqToMax | maxLeqL, maxLeqR, maxLeast)

Methods

leqToCmp :: Sing a -> Sing b -> IsTrue (a <= b) -> Either (a :~: b) (Compare a b :~: LT)

eqlCmpEQ :: Sing a -> Sing b -> (a :~: b) -> Compare a b :~: EQ

eqToRefl :: Sing a -> Sing b -> (Compare a b :~: EQ) -> a :~: b

flipCompare :: Sing a -> Sing b -> FlipOrdering (Compare a b) :~: Compare b a

ltToNeq :: Sing a -> Sing b -> (Compare a b :~: LT) -> (a :~: b) -> Void

leqNeqToLT :: Sing a -> Sing b -> IsTrue (a <= b) -> ((a :~: b) -> Void) -> Compare a b :~: LT

succLeqToLT :: Sing a -> Sing b -> IsTrue (S a <= b) -> Compare a b :~: LT

ltToLeq :: Sing a -> Sing b -> (Compare a b :~: LT) -> IsTrue (a <= b)

gtToLeq :: Sing a -> Sing b -> (Compare a b :~: GT) -> IsTrue (b <= a)

ltToSuccLeq :: Sing a -> Sing b -> (Compare a b :~: LT) -> IsTrue (Succ a <= b)

cmpZero :: Sing a -> Compare (Zero nat) (Succ a) :~: LT

leqToGT :: Sing a -> Sing b -> IsTrue (Succ b <= a) -> Compare a b :~: GT

cmpZero' :: Sing a -> Either (Compare (Zero nat) a :~: EQ) (Compare (Zero nat) a :~: LT)

zeroNoLT :: Sing a -> (Compare a (Zero nat) :~: LT) -> Void

ltRightPredSucc :: Sing a -> Sing b -> (Compare a b :~: LT) -> b :~: Succ (Pred b)

cmpSucc :: Sing n -> Sing m -> Compare n m :~: Compare (Succ n) (Succ m)

ltSucc :: Sing a -> Compare a (Succ a) :~: LT

cmpSuccStepR :: Sing n -> Sing m -> (Compare n m :~: LT) -> Compare n (Succ m) :~: LT

ltSuccLToLT :: Sing n -> Sing m -> (Compare (Succ n) m :~: LT) -> Compare n m :~: LT

leqToLT :: Sing a -> Sing b -> IsTrue (Succ a <= b) -> Compare a b :~: LT

leqZero :: Sing n -> IsTrue (Zero nat <= n)

leqSucc :: Sing n -> Sing m -> IsTrue (n <= m) -> IsTrue (Succ n <= Succ m)

fromLeqView :: LeqView n m -> IsTrue (n <= m)

leqViewRefl :: Sing n -> LeqView n n

viewLeq :: Sing n -> Sing m -> IsTrue (n <= m) -> LeqView n m

leqWitness :: Sing n -> Sing m -> IsTrue (n <= m) -> DiffNat n m

leqStep :: Sing n -> Sing m -> Sing l -> ((n + l) :~: m) -> IsTrue (n <= m)

leqNeqToSuccLeq :: Sing n -> Sing m -> IsTrue (n <= m) -> ((n :~: m) -> Void) -> IsTrue (Succ n <= m)

leqRefl :: Sing n -> IsTrue (n <= n)

leqSuccStepR :: Sing n -> Sing m -> IsTrue (n <= m) -> IsTrue (n <= Succ m)

leqSuccStepL :: Sing n -> Sing m -> IsTrue (Succ n <= m) -> IsTrue (n <= m)

leqReflexive :: Sing n -> Sing m -> (n :~: m) -> IsTrue (n <= m)

leqTrans :: Sing n -> Sing m -> Sing l -> IsTrue (n <= m) -> IsTrue (m <= l) -> IsTrue (n <= l)

leqAntisymm :: Sing n -> Sing m -> IsTrue (n <= m) -> IsTrue (m <= n) -> n :~: m

plusMonotone :: Sing n -> Sing m -> Sing l -> Sing k -> IsTrue (n <= m) -> IsTrue (l <= k) -> IsTrue ((n + l) <= (m + k))

leqZeroElim :: Sing n -> IsTrue (n <= Zero nat) -> n :~: Zero nat

plusMonotoneL :: Sing n -> Sing m -> Sing l -> IsTrue (n <= m) -> IsTrue ((n + l) <= (m + l))

plusMonotoneR :: Sing n -> Sing m -> Sing l -> IsTrue (m <= l) -> IsTrue ((n + m) <= (n + l))

plusLeqL :: Sing n -> Sing m -> IsTrue (n <= (n + m))

plusLeqR :: Sing n -> Sing m -> IsTrue (m <= (n + m))

plusCancelLeqR :: Sing n -> Sing m -> Sing l -> IsTrue ((n + l) <= (m + l)) -> IsTrue (n <= m)

plusCancelLeqL :: Sing n -> Sing m -> Sing l -> IsTrue ((n + m) <= (n + l)) -> IsTrue (m <= l)

succLeqZeroAbsurd :: Sing n -> IsTrue (S n <= Zero nat) -> Void

succLeqZeroAbsurd' :: Sing n -> (S n <= Zero nat) :~: False

succLeqAbsurd :: Sing n -> IsTrue (S n <= n) -> Void

succLeqAbsurd' :: Sing n -> (S n <= n) :~: False

notLeqToLeq :: (n <= m) ~ False => Sing n -> Sing m -> IsTrue (m <= n)

leqSucc' :: Sing n -> Sing m -> (n <= m) :~: (Succ n <= Succ m)

leqToMin :: Sing n -> Sing m -> IsTrue (n <= m) -> Min n m :~: n

geqToMin :: Sing n -> Sing m -> IsTrue (m <= n) -> Min n m :~: m

minComm :: Sing n -> Sing m -> Min n m :~: Min m n

minLeqL :: Sing n -> Sing m -> IsTrue (Min n m <= n)

minLeqR :: Sing n -> Sing m -> IsTrue (Min n m <= m)

minLargest :: Sing l -> Sing n -> Sing m -> IsTrue (l <= n) -> IsTrue (l <= m) -> IsTrue (l <= Min n m)

leqToMax :: Sing n -> Sing m -> IsTrue (n <= m) -> Max n m :~: m

geqToMax :: Sing n -> Sing m -> IsTrue (m <= n) -> Max n m :~: n

maxComm :: Sing n -> Sing m -> Max n m :~: Max m n

maxLeqR :: Sing n -> Sing m -> IsTrue (m <= Max n m)

maxLeqL :: Sing n -> Sing m -> IsTrue (n <= Max n m)

maxLeast :: Sing l -> Sing n -> Sing m -> IsTrue (n <= l) -> IsTrue (m <= l) -> IsTrue (Max n m <= l)

leqReversed :: Sing n -> Sing m -> (n <= m) :~: (m >= n)

lneqSuccLeq :: Sing n -> Sing m -> (n < m) :~: (Succ n <= m)

lneqReversed :: Sing n -> Sing m -> (n < m) :~: (m > n)

lneqToLT :: Sing n -> Sing m -> IsTrue (n < m) -> Compare n m :~: LT

ltToLneq :: Sing n -> Sing m -> (Compare n m :~: LT) -> IsTrue (n < m)

lneqZero :: Sing a -> IsTrue (Zero nat < Succ a)

lneqSucc :: Sing n -> IsTrue (n < Succ n)

succLneqSucc :: Sing n -> Sing m -> (n < m) :~: (Succ n < Succ m)

lneqRightPredSucc :: Sing n -> Sing m -> IsTrue (n < m) -> m :~: Succ (Pred m)

lneqSuccStepL :: Sing n -> Sing m -> IsTrue (Succ n < m) -> IsTrue (n < m)

lneqSuccStepR :: Sing n -> Sing m -> IsTrue (n < m) -> IsTrue (n < Succ m)

plusStrictMonotone :: Sing n -> Sing m -> Sing l -> Sing k -> IsTrue (n < m) -> IsTrue (l < k) -> IsTrue ((n + l) < (m + k))

maxZeroL :: Sing n -> Max (Zero nat) n :~: n

maxZeroR :: Sing n -> Max n (Zero nat) :~: n

minZeroL :: Sing n -> Min (Zero nat) n :~: Zero nat

minZeroR :: Sing n -> Min n (Zero nat) :~: Zero nat

minusSucc :: Sing n -> Sing m -> IsTrue (m <= n) -> (Succ n - m) :~: Succ (n - m)

lneqZeroAbsurd :: Sing n -> IsTrue (n < Zero nat) -> Void

minusPlus :: PeanoOrder nat => Sing n -> Sing m -> IsTrue (m <= n) -> ((n - m) + m) :~: n

(===) :: Equality eq => eq x y -> Reason eq y z -> eq x z

(=~=) :: r x y -> Sing y -> r x y

because :: Sing y -> eq x y -> Reason eq x y

coerce :: (a :=: b) -> f a -> f b

start :: Preorder eq => Sing a -> eq a a

withWitness :: IsTrue b -> ((b ~ True) -> r) -> r

data IsTrue (b :: Bool) where

Constructors

Witness :: IsTrue True

type family Flipped f a :: Nat -> Type where ...

Equations

Flipped f a = Flipped f a

type SNat (n :: Nat) = Sing n

type Sized' n a = Sized Seq n a

type Sized n a = Sized Vector n a

pattern Flipped :: Sized f n a -> Flipped f a n

toProxy :: a -> Proxy a

coerceLength :: (n :~: m) -> Sized f n a -> Sized f m a

sizedLength :: ListLike (f a) a => Sized f n a -> Sing n

padVecs :: forall a n m. a -> Sized' n a -> Sized' m a -> (SNat (Max n m), Sized' (Max n m) a, Sized' (Max n m) a)

sNatToInt :: SNat n -> Int

Orphan instances

Hashable a => Hashable (Seq a)

Instance details

Methods

hashWithSalt :: Int -> Seq a -> Int

hash :: Seq a -> Int