Algebra.Ring.Polynomial.Monomial

Documentation

type Monomial n = Sized' n Int

N-ary Monomial. IntMap contains degrees for each x_i- type Monomial (n :: Nat) = Sized n Int

newtype OrderedMonomial ordering n

A wrapper for monomials with a certain (monomial) order.

Constructors

OrderedMonomial

Fields

getMonomial :: Monomial n

class IsOrder (n :: Nat) (ordering :: Type) where

Class to lookup ordering from its (type-level) name.

Minimal complete definition

cmpMonomial

Methods

cmpMonomial :: Proxy ordering -> MonomialOrder n

class IsOrder n name => IsMonomialOrder n name

Class for Monomial orders.

type MonomialOrder n = Monomial n -> Monomial n -> Ordering

Monomial order (of degree n). This should satisfy following laws: (1) Totality: forall a, b (a < b || a == b || b < a) (2) Additivity: a b == a + c <= b + c (3) Non-negative: forall a, 0 <= a

type IsStrongMonomialOrder ord = Forall (IsMonomialOrder' ord)

Monomial ordering which can do with monomials of arbitrary large arity.

isRelativelyPrime :: OrderedMonomial ord n -> OrderedMonomial ord n -> Bool

totalDegree :: OrderedMonomial ord n -> Int

data ProductOrder (n :: Nat) (m :: Nat) (a :: Type) (b :: Type) where

Constructors

ProductOrder :: Sing n -> Sing m -> ord -> ord' -> ProductOrder n m ord ord'

productOrder :: forall ord ord' n m. (IsOrder n ord, IsOrder m ord', KnownNat n, KnownNat m) => Proxy (ProductOrder n m ord ord') -> MonomialOrder (n + m)

productOrder' :: forall n ord ord' m. (IsOrder n ord, IsOrder m ord') => SNat n -> SNat m -> ord -> ord' -> MonomialOrder (n + m)

type WeightProxy (v :: [Nat]) = SList v

data WeightOrder (v :: [Nat]) (ord :: Type) where

Constructors

WeightOrder :: SList (v :: [Nat]) -> Proxy ord -> WeightOrder v ord

gcdMonomial :: OrderedMonomial ord n -> OrderedMonomial ord n -> OrderedMonomial ord n

divs :: OrderedMonomial ord n -> OrderedMonomial ord n -> Bool

isPowerOf :: KnownNat n => OrderedMonomial ord n -> OrderedMonomial ord n -> Bool

tryDiv :: Field r => (r, OrderedMonomial ord n) -> (r, OrderedMonomial ord n) -> (r, OrderedMonomial ord n)

lcmMonomial :: OrderedMonomial ord n -> OrderedMonomial ord n -> OrderedMonomial ord n

data Lex

Lexicographical order

Constructors

Lex

class (IsMonomialOrder n ord, KnownNat n) => EliminationType n m ord

Monomial order which can be use to calculate n-th elimination ideal of m-ary polynomial. This should judge monomial to be bigger if it contains variables to eliminate.

type EliminationOrder n m = ProductOrder n m Grevlex Grevlex

type WeightedEliminationOrder (n :: Nat) (ord :: Type) = WeightOrder (Replicate n 1) ord

eliminationOrder :: SNat n -> SNat m -> EliminationOrder n m

weightedEliminationOrder :: SNat n -> WeightedEliminationOrder n Grevlex

lex :: MonomialOrder n

Lexicographical order. This *is* a monomial order.

revlex :: MonomialOrder n

Reversed lexicographical order. This is *not* a monomial order.

graded :: MonomialOrder n -> MonomialOrder n

Convert ordering into graded one.

grlex :: MonomialOrder n

Graded lexicographical order. This *is* a monomial order.

grevlex :: MonomialOrder n

Graded reversed lexicographical order. This *is* a monomial order.

weightOrder :: forall n ns ord. (KnownNat n, IsOrder n ord, SingI ns) => Proxy (WeightOrder ns ord) -> MonomialOrder n

data Grevlex

Graded reversed lexicographical order. Same as Graded Revlex.

Constructors

Grevlex

fromList :: SNat n -> [Int] -> Monomial n

convert NAry list into Monomial.

data Revlex

Reversed lexicographical order

Constructors

Revlex

data Grlex

Graded lexicographical order. Same as Graded Lex.

Constructors

Grlex

newtype Graded ord

Graded order from another monomial order.

Constructors

Graded ord

castMonomial :: KnownNat m => OrderedMonomial o n -> OrderedMonomial o' m

scastMonomial :: SNat m -> OrderedMonomial o n -> OrderedMonomial o m

varMonom :: SNat n -> Ordinal n -> Monomial n

changeMonomialOrder :: o' -> OrderedMonomial ord n -> OrderedMonomial o' n

changeMonomialOrderProxy :: Proxy o' -> OrderedMonomial ord n -> OrderedMonomial o' n

sOnes :: Sing n -> Sing (Replicate n 1)

withStrongMonomialOrder :: forall ord n r proxy (proxy' :: Nat -> Type). IsStrongMonomialOrder ord => proxy ord -> proxy' n -> (IsMonomialOrder n ord => r) -> r

cmpAnyMonomial :: IsStrongMonomialOrder ord => Proxy ord -> Monomial n -> Monomial m -> Ordering

Comparing monomials with different arity, padding with 0 at bottom of the shorter monomial to make the length equal.

orderMonomial :: proxy ord -> Monomial n -> OrderedMonomial ord n

Orphan instances

Ord (Monomial n)
For simplicity, we choose grevlex for the default monomial ordering (for the sake of efficiency).

Instance details

Methods

compare :: Monomial n -> Monomial n -> Ordering

(<) :: Monomial n -> Monomial n -> Bool

(<=) :: Monomial n -> Monomial n -> Bool

(>) :: Monomial n -> Monomial n -> Bool

(>=) :: Monomial n -> Monomial n -> Bool

max :: Monomial n -> Monomial n -> Monomial n

min :: Monomial n -> Monomial n -> Monomial n

KnownNat n => Unital (Monomial n)

Instance details

Methods

one :: Monomial n

pow :: Monomial n -> Natural -> Monomial n

productWith :: Foldable f => (a -> Monomial n) -> f a -> Monomial n

Multiplicative (Monomial n)

Instance details

Methods

(*) :: Monomial n -> Monomial n -> Monomial n

pow1p :: Monomial n -> Natural -> Monomial n

productWith1 :: Foldable1 f => (a -> Monomial n) -> f a -> Monomial n