Algebra.Internal

Documentation

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall k (a :: k). a :~: a

withRefl :: forall k (a :: k) (b :: k) r. (a :~: b) -> (a ~ b => r) -> r

asProxyTypeOf :: a -> proxy a -> a

asProxyTypeOf is a type-restricted version of const. It is usually used as an infix operator, and its typing forces its first argument (which is usually overloaded) to have the same type as the tag of the second.

>>> import Data.Word
>>> :type asProxyTypeOf 123 (Proxy :: Proxy Word8)
asProxyTypeOf 123 (Proxy :: Proxy Word8) :: Word8

Note the lower-case proxy in the definition. This allows any type constructor with just one argument to be passed to the function, for example we could also write

>>> import Data.Word
>>> :type asProxyTypeOf 123 (Just (undefined :: Word8))
asProxyTypeOf 123 (Just (undefined :: Word8)) :: Word8

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

data KProxy t

A concrete, promotable proxy type, for use at the kind level. There are no instances for this because it is intended at the kind level only

Constructors

KProxy

class KnownNat (n :: Nat)

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: base-4.7.0.0

Minimal complete definition

natSing

data Nat

(Kind) This is the kind of type-level natural numbers.

type family (a :: Nat) + (b :: Nat) :: Nat where ... infixl 6

Addition of type-level naturals.

Since: base-4.7.0.0

type family (a :: Nat) * (b :: Nat) :: Nat where ... infixl 7

Multiplication of type-level naturals.

Since: base-4.7.0.0

type family (a :: Nat) ^ (b :: Nat) :: Nat where ... infixr 8

Exponentiation of type-level naturals.

Since: base-4.7.0.0

type family (a :: Nat) <=? (b :: Nat) :: Bool where ... infix 4

Comparison of type-level naturals, as a function. NOTE: The functionality for this function should be subsumed by CmpNat, so this might go away in the future. Please let us know, if you encounter discrepancies between the two.

type family (a :: Nat) - (b :: Nat) :: Nat where ... infixl 6

Subtraction of type-level naturals.

Since: base-4.7.0.0

type family CmpNat (a :: Nat) (b :: Nat) :: Ordering where ...

Comparison of type-level naturals, as a function.

Since: base-4.7.0.0

type family Div (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Division (round down) of natural numbers. Div x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

type family Mod (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Modulus of natural numbers. Mod x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

type family Log2 (a :: Nat) :: Nat where ...

Log base 2 (round down) of natural numbers. Log 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

someNatVal :: Natural -> SomeNat

Convert an integer into an unknown type-level natural.

Since: base-4.10.0.0

natVal' :: forall (n :: Nat). KnownNat n => Proxy# n -> Natural

Since: base-4.10.0.0

natVal :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Natural

Since: base-4.10.0.0

data SomeNat

This type represents unknown type-level natural numbers.

Since: base-4.10.0.0

Constructors

KnownNat n => SomeNat (Proxy n)

type (<=) (x :: Nat) (y :: Nat) = (x <=? y) ~ 'True infix 4

Comparison of type-level naturals, as a constraint.

Since: base-4.7.0.0

withSingI :: forall k (n :: k) r. Sing n -> (SingI n => r) -> r

type family Demote k = (r :: Type) | r -> k

type family Sing :: k -> Type

class SingI (a :: k) where

Methods

sing :: Sing a

class SingKind k where

Associated Types

type Demote k = (r :: Type) | r -> k

Methods

fromSing :: forall (a :: k). Sing a -> Demote k

toSing :: Demote k -> SomeSing k

data SomeSing k where

Constructors

SomeSing :: forall k (a :: k). Sing a -> SomeSing k

pattern (:<) :: (Dom f a, KnownNat n, CFreeMonoid f) => (n ~ (1 + n1), KnownNat n1) => a -> Sized f n1 a -> Sized f n a

pattern (:>) :: (Dom f a, KnownNat n, CFreeMonoid f) => (n ~ (n1 + 1), KnownNat n1) => Sized f n1 a -> a -> Sized f n a

pattern Nil :: forall f n a. (KnownNat n, CFreeMonoid f, Dom f a) => n ~ 0 => Sized f n a

generate :: forall (f :: Type -> Type) (n :: Nat) a. (CFreeMonoid f, Dom f a) => SNat n -> (Ordinal n -> a) -> Sized f n a

sIndex :: forall (f :: Type -> Type) (n :: Nat) c. (CFoldable f, Dom f c) => Ordinal n -> Sized f n c -> c

singleton :: forall (f :: Type -> Type) a. (CPointed f, Dom f a) => a -> Sized f 1 a

unsafeFromList :: forall (f :: Type -> Type) (n :: Nat) a. (CFreeMonoid f, Dom f a) => SNat n -> [a] -> Sized f n a

unsafeFromList' :: forall (f :: Type -> Type) (n :: Nat) a. (KnownNat n, CFreeMonoid f, Dom f a) => [a] -> Sized f n a

zipWithSame :: forall (f :: Type -> Type) (n :: Nat) a b c. (Dom f a, CZip f, Dom f b, CFreeMonoid f, Dom f c) => (a -> b -> c) -> Sized f n a -> Sized f n b -> Sized f n c

viewNat :: forall (n :: Nat). SNat n -> ZeroOrSucc n

type Pred (n :: Nat) = n - 1

data SNat (n :: Nat) where

Bundled Patterns

pattern Zero :: () => n ~ 0 => SNat n

pattern Succ :: forall n n1. () => n ~ Succ n1 => SNat n1 -> SNat n

type (<) (n :: Nat) (m :: Nat) = (n <? m) ~ 'True

type Succ (n :: Nat) = n + 1

type (-.) (n :: Nat) (m :: Nat) = Subt n m (m <=? n)

type S (n :: Nat) = Succ n

type Min (n :: Nat) (m :: Nat) = MinAux (n <=? m) n m

withKnownNat :: forall (n :: Nat) r. SNat n -> (KnownNat n => r) -> r

toNatural :: forall (n :: Nat). SNat n -> Natural

sNatP :: forall (n :: Nat) pxy. KnownNat n => pxy n -> SNat n

snat :: QuasiQuoter

toSomeSNat :: Natural -> SomeSNat

withSNat :: Natural -> (forall (n :: Nat). KnownNat n => SNat n -> r) -> r

(%*) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n * m)

(%+) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n + m)

(%-) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n - m)

(%<=?) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SBool (n <=? m)

(%^) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n ^ m)

sCmpNat :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SOrdering (CmpNat n m)

sCompare :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SOrdering (CmpNat n m)

sDiv :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (Div n m)

sFlipOrdering :: forall (ord :: Ordering). SOrdering ord -> SOrdering (FlipOrdering ord)

sLog2 :: forall (n :: Nat). SNat n -> SNat (Log2 n)

sMod :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (Mod n m)

sNat :: forall (n :: Nat). KnownNat n => SNat n

sOne :: SNat 1

sPred :: forall (n :: Nat). SNat n -> SNat (Pred n)

sS :: forall (n :: Nat). SNat n -> SNat (Succ n)

sSucc :: forall (n :: Nat). SNat n -> SNat (Succ n)

sZero :: SNat 0

induction :: forall p (k :: Nat). p 0 -> (forall (n :: Nat). SNat n -> p n -> p (S n)) -> SNat k -> p k

zeroOrSucc :: forall (n :: Nat). SNat n -> ZeroOrSucc n

(%-.) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n -. m)

(%<?) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SBool (n <? m)

(%>=?) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SBool (n >=? m)

(%>?) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SBool (n >? m)

sMax :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (Max n m)

sMin :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (Min n m)

data SomeSNat where

Constructors

SomeSNat :: forall (n :: Nat). KnownNat n => SNat n -> SomeSNat

type family (a :: k) === (b :: k) :: Bool where ...

Equations

(a :: k) === (a :: k) = 'True

(_1 :: k) === (_2 :: k) = 'False

data Equality (n :: k) (m :: k) where

Constructors

Equal :: forall k (n :: k). (n == n) ~ 'True => Equality n n

NonEqual :: forall k (n :: k) (m :: k). ((n === m) ~ 'False, (n == m) ~ 'False) => Equality n m

type family FlipOrdering (ord :: Ordering) :: Ordering where ...

Equations

FlipOrdering 'LT = 'GT

FlipOrdering 'GT = 'LT

FlipOrdering 'EQ = 'EQ

type One = 1

data SBool (b :: Bool) where

Constructors

SFalse :: SBool 'False

STrue :: SBool 'True

data SOrdering (ord :: Ordering) where

Constructors

SLT :: SOrdering 'LT

SEQ :: SOrdering 'EQ

SGT :: SOrdering 'GT

type Zero = 0

data ZeroOrSucc (n :: Nat) where

Constructors

IsZero :: ZeroOrSucc 0

IsSucc :: forall (n1 :: Nat). SNat n1 -> ZeroOrSucc (n1 + 1)

type (<?) (n :: Nat) (m :: Nat) = (n + 1) <=? m

type (>) (n :: Nat) (m :: Nat) = (n >? m) ~ 'True

type (>=) (n :: Nat) (m :: Nat) = (n >=? m) ~ 'True

type (>=?) (n :: Nat) (m :: Nat) = m <=? n

type (>?) (n :: Nat) (m :: Nat) = m <? n

type Max (n :: Nat) (m :: Nat) = MaxAux (n >=? m) n m

data IsTrue (b :: Bool) where

Constructors

Witness :: IsTrue 'True

withWitness :: forall (b :: Bool) r. IsTrue b -> (b ~ 'True => r) -> r

(===) :: forall k eq (x :: k) (y :: k) (z :: k). Equality eq => eq x y -> Reason eq y z -> eq x z

(=~=) :: forall k1 k2 r (x :: k1) (y :: k2) proxy. r x y -> proxy y -> r x y

because :: forall k1 k2 proxy (y :: k1) eq (x :: k2). proxy y -> eq x y -> Reason eq x y

coerce :: forall k (a :: k) (b :: k) f. (a :=: b) -> f a -> f b

start :: forall k eq proxy (a :: k). Preorder eq => proxy a -> eq a a

type family Flipped f a :: Nat -> Type where ...

Equations

Flipped f a = Flipped f a

type Sized' n a = Sized Seq n a

type Sized n a = Sized Vector n a

type USized n a = Sized Vector n a

pattern Flipped :: Sized f n a -> Flipped f a n

toProxy :: a -> Proxy a

sNatToSingleton :: SNat n -> Sing n

singToSNat :: Sing n -> SNat n

coerceLength :: (n :~: m) -> Sized f n a -> Sized f m a

sizedLength :: (KnownNat n, DomC f a) => Sized f n a -> SNat n

padVecs :: forall a n m. (Unbox a, KnownNat n, KnownNat m) => a -> USized n a -> USized m a -> (SNat (Max n m), USized (Max n m) a, USized (Max n m) a)

sNatToInt :: SNat n -> Int

Orphan instances

Hashable a => Hashable (Seq a)

Instance details

Methods

hashWithSalt :: Int -> Seq a -> Int

hash :: Seq a -> Int